Move system info related endpoints to separate service part
